+++ DTKit/view/DTGridTableView/DTGridTableViewCell.m
@@ -20,7 +20,7 @@
 
 - (id)initWithStyle:(UITableViewCellStyle)style reuseIdentifier:(NSString *)reuseIdentifier
            delegate:(id<DTGridTableViewCellDelegate>)delegate
-          rowNumber:(int)number rowHeight:(float)height indexPath:(NSIndexPath *)indexPath
+          rowNumber:(NSInteger)number rowHeight:(CGFloat)height indexPath:(NSIndexPath *)indexPath
 {
     self = [super initWithStyle:style reuseIdentifier:reuseIdentifier];
     if (self) {
